backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
PyYAML==6.0.2
httpx==0.27.0
pydantic==2.8.2
python-multipart==0.0.9
